+++ channel-poll1/src/main/shared/Statement.scala
@@ -1,5 +1,5 @@
 package main.shared
 
-class Statement (val message: String, val userID: String, val userName: String, val screenName: String, val pictureURL: String, val creationDate: String, val ID: Int){
+class Statement (val message: String, val userID: String, val userName: String, val screenName: String, val pictureURL: String, val creationDate: String, val ID: Long){
 
 }
